+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Preserving ProSMART parameters when cloning Servalcat jobs
 - Changed coot_rsr_morph and coot_add_waters to use CHAPI
 - Fix to Servalcat USE_WORK_IN_EST
+- Fix to user selected columns in split MTZ task
 
 ## [2.4.3] - 2025-10-24